Move mocks to separate files so they are re-usable
- Rather than manually defining each mock in every test file, move them
to mock files that can be loaded when required. This way mocks can be
used amongst many tests easily.
