+++ src/__mocks__/vue-i18n.ts
@@ -1,0 +1,9 @@
+export function useI18n() {
+  return {
+    useI18n: jest.fn().mockImplementation(() => {
+      return {
+        t: jest.fn().mockImplementation()
+      };
+    })
+  };
+}
