+++ src/composables/__mocks__/useEthereumTxType.ts
@@ -1,0 +1,10 @@
+export default function useEthereumTxType() {
+  return {
+    EthereumTxType: {
+      EIP1559: 'EIP1559'
+    },
+    ethereumTxType: {
+      value: 'EIP1559'
+    }
+  };
+}
