+++ src/composables/__mocks__/useEthers.ts
@@ -1,0 +1,5 @@
+export default function useEthers() {
+  return {
+    txListener: jest.fn().mockImplementation()
+  };
+}
