+++ src/composables/__mocks__/useTransactions.ts
@@ -1,0 +1,5 @@
+export default function useTransactions() {
+  return {
+    addTransaction: jest.fn().mockImplementation()
+  };
+}
