+++ src/locales/__mocks__/index.ts
@@ -1,0 +1,1 @@
+export default {};
